aggregateur/requirements.txt
GitPython==2.1.11
git-url-parse==1.2.2
semver==2.8.1
tableschema==1.4.1
PyYAML==5.1
mailjet-rest==1.3.3
python-frontmatter==0.4.5
table_schema_to_markdown==0.3.1
toml==0.10.0
